Using the cluster's container-runtime in the worker pod.
There is a new cri-client available to interact with the cri-api which
defines the cluster's container-runtime interface.

It allows us to use the cluster's container-runtime for pulling images
instead of manually pulling them in the worker pod.

The benefits of using the cluster's container-runtime is mainly to
leverage kubelet's images cache on the node and reduce network
bandwidth.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>